added thread lock and rmoved for to delete msg

+++ bot/cogs/events/message.py
@@ -4,10 +4,13 @@ import discord
 from discord import app_commands, Interaction
 from discord.ext import commands, tasks
 
+from threading import Lock
+
 class Messages(commands.Cog):
     def __init__(self, client:Bot):
         self.client:Bot = client
         self.clear_list = []
+        self.clear_list_mutex = Lock()
         super().__init__()
 
 #   COMMANDS
@@ -17,10 +20,11 @@ class Messages(commands.Cog):
         channel = interaction.channel
         user = interaction.user
         try:
-            async for message in channel.history(limit=limit):
-                self.clear_list.append(message)
-            print(f"Message sent to {channel.name}")
-            await interaction.response.send_message(content="msgs added in list", ephemeral=True, delete_after=5)
+            with self.clear_list_mutex:
+                async for message in channel.history(limit=limit):
+                    self.clear_list.append(message)
+                print(f"Message sent to {channel.name}")
+                await interaction.response.send_message(content="msgs added in list", ephemeral=True, delete_after=5)
         except Exception as e:
             print(f"Error add msg in clear list {channel.name}: {e}")
             await interaction.response.send_message(content=f"Error add msg in clear list {channel.name}: {e}",
@@ -29,7 +33,6 @@ class Messages(commands.Cog):
 #   EVENTS
     @commands.Cog.listener()
     async def on_ready(self):
-        await self.client.tree.sync()
         self.start_loop()
         print("started Messages")
     
@@ -38,13 +41,15 @@ class Messages(commands.Cog):
         pass
 
 #   TASKS
-    @tasks.loop(seconds=1, name="clear_list_loop")
+    @tasks.loop(seconds=0.3, name="clear_list_loop")
     async def clear_list_loop(self):
         try:
-            for msg in self.clear_list:
-                msg:discord.Message
+            with self.clear_list_mutex:
+                if not self.clear_list:
+                    return 
+                msg:discord.Message = self.clear_list[0]
                 await msg.delete()
-                self.clear_list.remove(msg)
+                self.clear_list.pop(0)
         except Exception as e:
             print(f"Error clear_list_loop message to {e}")
